Revert "made hero section responsive and created mobile navbar"
